Restructuring preparing to work on feature
