+++ WindowsSetupAssistant/WindowsSetupAssistant.Core/Models/IInstallables/Interfaces/IInstallable.cs
