+++ documentation/img/getting-started/MainWindow.png
