+++ documentation/img/getting-started/bat_file_windows.png
